clear out test repository before running tests
We were running into problems with spec/features/create_document_spec.rb
where the 'Add new Work' button wasn't appearing. After a lot of poking
around, it looked like the default admin set was present, but its
permission_template was not, likely because the former is stored in the
repository and the latter is stored in the database, which was being
cleaned out after each test. So the default admin set wasn't being
created each test, but its permission template was being cleared out,
and because of this, users weren't able to create new works.