add support for creating nested_attributes for based_near
We want to create items on import that have RDF mappings for
location (where present). Previously, when we tried passing
RDF::URI objects in the mappers, we were met with Job errors
that stated that the URI objects couldn't be serialized. By
storing these values in `based_near_attributes` I think we
can get around this?