add local_controlled_vocabulary_input type
we can't use the hyrax +controlled_vocabulary_input+
for local vocabularies, since that input requires
+ActiveTriples::Resource+ classed-properties, which
is a little excessive for our purposes.

i have a personal preference to the ui of the LinkedData
hydra-editor field type over the plain jquery-ui Default
input, so this hydra-editor type builds inputs that
LinkedData is expecting.