remove the option to alter ui colors
this strips out the 'Appearance' menu from the admin dashboard
that allowed users to choose colors for the site. this will
allow us to have the same look on each deploy (without having
to write an initializer to set the colors we want) and allow
us to theme the site a bit better (the way Hyrax applies the
custom styles is to add a <style> tag to the <head> of the layout,
which overrides any styles defined before).