looks like we'll need :facetable after all
unless we want to go in and do some fiddling on the hyrax renderers,
we'll need to index fields that we want the ability to facet on from
within the item-show page as :facetable, rather than :symbol. to
ensure that the values are stored, we'll also index these as :stored_searchable.