let blacklight determine the field label after configuration
this is a bandaid for a problem i don't completely understand.
all of a sudden, blacklight facet/field labels were displaying
the I18n 'translation missing' error instead of their values.
from what i can surmise, this is because the CatalogController
is being loaded _before_ the locales are. i came to this conclusion
after inserting a +puts "RESOURCE_TYPE #{I18n.t('blacklight.search.fields.resource_type')}"+
call in the +configure_blacklight+ block and, after starting
the server, saw "RESOURCE_TYPE translation missing ..." text
on the console.

i _think_ it may trace back to the Bootsnap dependency
(commenting out the 'bootsnap/setup' call in boot.rb fixed
the problem, if i recall correctly), but updating the
gem didn't do anything (i've reverted the commit where
i did that in the interest of having a dependency update
sprint after the ui one).

so hopefully this will fix the problem. i'm not at all into
the concept of having to provide translations for all of the
solrized edge cases ('publisher' and 'publisher_sim' and 'publisher_tesim')
that are used around the app, but we might be able to remediate
that by a Blacklight field configuration... maybe?

ugh, this is long. keeping my fingers crossed.