[circle-ci] try splitting tests by name rather than timings
not 100% sure what's going wrong, but i'm getting
failing specs on circle-ci that are passing locally.
since it always happens on the fourth batch of split
tests, perhaps splitting by timing is the culprit?

probably going to want to squash these if they get
out of hand trying to debug.