load overrides in after_initialize block (#214)
changes how overrides are loaded to prevent weird/random conflicts with the I18n gem.
previously, the overloads we causing "translation missing" messages all over the place
for engine locales. 

this performs our local hyrax overrides in the `after_initialize` application callback,
and then pushes the lambda to the end of the `to_prepare` array of blocks so that
it's the last performed then.

also included is a fix for the `create_publication` feature spec that was failing in
chrome v74 (and beyond?).