store extracted-text at the work level (#298)
now stores (as well as indexes) full-text content at the publication level,
rather than at the file_set. this requires us removing the hyrax join query
(which searches file_sets as well as works), but allows us to now use the
blacklight-advanced-search plugin properly. 

this also gives us the ability to display contextual match highlighting within
search results. as i'm not entirely sure of the storage/performance
consequences/costs, i've set-this up as a toggle-able feature (within
`/admin/features`) so that we can continue to test it before enabling
it permanently. 

squashed commits:

* store/index full-text content at the publication level
* don't index full-text content at the file_set level
* use extracted_text_tsimv field for searches
* display solr highlight hits for full-text on search results
* add/update specs
* use fastVector for highlighting text
* update labels for advanced-search form
* revert d3bdec7 temporarily
  * this'll allow us to test this branch w/o having to reingest everything that's live on stage
* hard-code labels for search fields for now
* reinstate d3bdec7 (/ldr base for prod fedora)
* remove all_text_timv solr copy field
* make displaying highlighted search results a toggleable feature