fix recurring characterization encoding problems
so it turns out that the utf-8 encoding problems were
stemming from an out-of-date version of FITSServlet
that we're stuck using (1.1.3). v1.2.0 explicitly
sets the response encoding to utf-8, but when i've tried
updating the version, i get exceptions thrown that
i can't track down. so instead, we'll convert the
encoding as low in the characterization stack as we can
in `Hydra::FileCharacterization::Characterizers::FitsServlet`.

undoes the work done in #317 by returning to the stock
Hyrax `CharacterizationJob`, but prefers the fits_servlet
tool when the FITS_SERVLET_URL environment variable is present.