revert single_use_keys table schema
converts snake_case keys to their original camelCase values:
- download_key -> downloadKey
- item_id -> itemId

this migration was added when we tried upgrading to hyrax@3
and was accidentally committed. when we felt the upgrade
was coming soon, the postgres tables were upgraded manually.
but as it's been quite a few months since any progress has
been made on the upgrade, we'll revert back to the hyrax@2
schema and RE-update the schema when we finally upgrade.